Moved styles around, fixed merged conflict
