+++ frontend/src/App.css
@@ -2,3 +2,7 @@
     display: flex;
     padding: 0;
 }
+
+.pt-md{
+    padding-top: 5em;
+}
