--- frontend/src/components/YesButton.js
@@ -1,14 +1,0 @@
-import react from "react";
-import '../styles/Buttons.css'
-
-function YesButton(){
-    var ButtonText = "YES";
-
-    return(
-        <button className="YesButton">
-            <div className="buttonText">{ButtonText}</div>
-        </button>
-    )
-}
-
-export default YesButton;
