+++ frontend/src/styles/Features.css
@@ -1,8 +1,4 @@
-<<<<<<< HEAD
-.container {
-=======
 .features-container {
->>>>>>> main
     display: flex;
     flex-direction: column;
     align-items: center;
@@ -12,12 +8,4 @@
     margin: 0 20px;
     list-style: none;
     padding: 0;
-<<<<<<< HEAD
-}
-
-.center {
-    display: flex;
-    padding: 0;
-=======
->>>>>>> main
 }
